*updated otitemeditor to "0.3.2", supports the official 8.20 client *fixed loads of warings *changed from expat to libxml2. *Updated project to MSVC 8 *Added "Copy Item" option

+++ client/item_sprite.hpp
@@ -121,8 +121,8 @@ public:
 	void unloadDat();
 	
 	unsigned short getSpriteDump(unsigned short id, char *dump);
-	unsigned short getSpriteMapSize() { return sprite.size();}
-	unsigned short getItemMapSize() { return item.size();}
+	size_t getSpriteMapSize() { return sprite.size();}
+	size_t getItemMapSize() { return item.size();}
 	unsigned short getItemSpriteID(unsigned short item_id, unsigned short frame );
 	
 	InternalSprite getSpriteInternalFormat(unsigned short item_id, unsigned short frame);
